requirements.txt
openai
chromadb
langchain
unstructured
pyyaml
streamlit
